fixed albums page ready to css aplication
