+++ client/src/components/AlbumList/AlbumListItem.tsx
@@ -1,5 +1,11 @@
-import { useEffect, useState } from "react";
+import { useRef, useState } from "react";
 import "./AlbumListItem.css";
+
+type Song = {
+  id: number;
+  title: string;
+  audioSrc: string;
+};
 
 type Album = {
   id: number;
@@ -9,17 +15,10 @@ type Album = {
   songs: Song[];
 };
 
-type Song = {
-  id: number;
-  title: string;
-  audioSrc: string;
-};
-
 function AlbumListItem({ album }: { album: Album }) {
   const [visibleSongs, setVisibleSongs] = useState(10);
-  const [currentAudio, setCurrentAudio] = useState<HTMLAudioElement | null>(
-    null,
-  );
+  const [isPlaying, setIsPlaying] = useState<number | null>(null);
+  const audioRefs = useRef<(HTMLAudioElement | null)[]>([]);
 
   const showMoreSongs = () => {
     setVisibleSongs((prev) =>
@@ -31,65 +30,69 @@ function AlbumListItem({ album }: { album: Album }) {
     setVisibleSongs(10);
   };
 
-  const playAudio = (audioSrc: string) => {
-    if (currentAudio) {
-      currentAudio.pause();
+  const handlePlay = (songId: number) => {
+    const audioElement = audioRefs.current[songId];
+    if (audioElement) {
+      if (audioElement.paused) {
+        audioElement.play();
+        setIsPlaying(songId);
+      } else {
+        audioElement.pause();
+        setIsPlaying(null);
+      }
     }
-
-    const newAudio = new Audio(audioSrc);
-    setCurrentAudio(newAudio);
-    newAudio.play();
   };
 
-  useEffect(() => {
-    return () => {
-      if (currentAudio) {
-        currentAudio.pause();
-      }
-    };
-  }, [currentAudio]);
+  const handleAudioRef = (el: HTMLAudioElement | null, songId: number) => {
+    audioRefs.current[songId] = el;
+  };
 
   return (
-    <div className="album-list-item" key={album.id}>
+    <div className="album-list-item">
       <div className="album-info">
         <img className="album-img" src={album.imgSrc} alt={album.title} />
         <h2 className="album-title">{album.title}</h2>
         <p className="album-description">{album.description}</p>
       </div>
+
       <div className="album-song-list">
         {album.songs.slice(0, visibleSongs).map((song, index) => (
           <div className="album-song-item" key={song.id}>
             <span className="album-song-number">{index + 1}</span>
             <p className="album-song-title">{song.title}</p>
-            {/* Bouton personnalisé pour jouer la chanson */}
             <button
               type="button"
-              className="album-play-button"
-              onClick={() => playAudio(song.audioSrc)}
+              className="album-play-btn"
+              onClick={() => handlePlay(song.id)}
             >
-              Play
+              {isPlaying === song.id ? "Pause" : "Play"}
             </button>
+            <audio
+              ref={(el) => handleAudioRef(el, song.id)}
+              src={song.audioSrc}
+              preload="metadata"
+            >
+              <track kind="captions" />
+            </audio>
           </div>
         ))}
-        <div className="album-song-controls">
-          {visibleSongs < album.songs.length ? (
-            <button
-              type="button"
-              className="album-show-more"
-              onClick={showMoreSongs}
-            >
-              Afficher plus
-            </button>
-          ) : (
-            <button
-              type="button"
-              className="album-show-less"
-              onClick={showLessSongs}
-            >
-              Afficher moins
-            </button>
-          )}
-        </div>
+        {visibleSongs < album.songs.length ? (
+          <button
+            type="button"
+            className="album-show-more"
+            onClick={showMoreSongs}
+          >
+            Afficher plus
+          </button>
+        ) : (
+          <button
+            type="button"
+            className="album-show-less"
+            onClick={showLessSongs}
+          >
+            Afficher moins
+          </button>
+        )}
       </div>
     </div>
   );
